refactor: [FX-2125] update InputAdornment
Currently, Icon and Limit adornments (which are defined in the Input
component's file) have classes for large and medium variants. This is
not helpful, as other variants of Input component also have Icon and
Limit adornments and we do not want to repeat ourselves. Thus, we
add large variant to InputAdornment instead, on which other Input
variants depend on.